feat: report skypilot status to discord channel (#2081)
This PR refactors the Discord webhook functionality into a reusable
utility module and adds Discord notifications for SkyPilot job launches.

## Changes

### Discord Utility Refactoring
- Moved Discord webhook functionality from
`.github/actions/discord-webhook/discord_webhook.py` to
`common/src/metta/common/util/discord.py`
- The GitHub Action script now imports from the common package instead
of duplicating the code
- Core functionality includes:
  - Automatic message splitting for Discord's 2000 character limit
  - Smart splitting at natural boundaries (paragraphs, lines, words)
  - Mention sanitization to prevent @everyone/@here pings
  - Rate limiting protection between messages

### SkyPilot Discord Notifications
- Added Discord webhook URL propagation through the launch pipeline:
  - GitHub workflow → GitHub action → launch script → SkyPilot config
- Jobs now post a notification to #eng-velocity when launched,
including:
  - Repository name
  - Git reference (branch/tag/commit)
  - Run ID
  - Launch timestamp
- Notifications fail gracefully without blocking job execution

### Other Changes
- Updated newsletter generation schedule from M/W/F to T/F
- Switched Discord webhook from `DISCORD_WEBHOOK_URL` to
`DISCORD_UPDATES_HOOK` secret

## Future Work

As noted in the notification message, this is currently testing phase -
we plan to restrict notifications to failed CI multinode runs only.


[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211039869118913)

---------
