+++ .github/workflows/generate-newsletter.yml
@@ -2,7 +2,7 @@ name: Generate Newsletter
 
 on:
   schedule:
-    - cron: "0 1 * * 2,4,6" # M/W/F at 1 AM GMT (5 PM PST)
+    - cron: "0 1 * * 3,6" # T/F at 1 AM GMT (5 PM PST)
   workflow_dispatch:
     inputs:
       days_to_scan:
@@ -220,7 +220,7 @@ jobs:
         if: ${{ !inputs.skip_discord }}
         uses: ./.github/actions/discord-webhook
         with:
-          webhook-url: ${{ secrets.DISCORD_WEBHOOK_URL }}
+          webhook-url: ${{ secrets.DISCORD_UPDATES_HOOK }}
           content-file: discord_summary_output.txt
 
       - name: Save PR Summaries Cache
