+++ devops/skypilot/launch.py
@@ -113,6 +113,10 @@ def main():
     parser.add_argument(
         "--github-pat", type=str, default=None, help="GitHub PAT token for posting status updates (repo scope)"
     )
+    parser.add_argument(
+        "--discord-webhook-url", type=str, default=None, help="Discord webhook URL for status update channel"
+    )
+
     (args, cmd_args) = parser.parse_known_args(filtered_args)
 
     if run_id is None:
@@ -153,6 +157,7 @@ def main():
         METTA_GIT_REF=commit_hash,
         HEARTBEAT_TIMEOUT=args.heartbeat_timeout_seconds,
         GITHUB_PAT=args.github_pat,
+        DISCORD_WEBHOOK_URL=args.discord_webhook_url,
     )
 
     env_updates = {k: v for k, v in env_updates.items() if v is not None}
